ECDC-4485: Update QPlot to Qt6 (#28)
Qt5 is frozen for develop, so going with Qt6 is properly the right path if new improvements and bug fixes needs to be addressed to the `QPlot` component. This PR addresses this

- Fix calls calls to deprecated Qt5 functionality in QPlot source files  
- Update CMake files to use Qt6
- Add build instructions to the README.md file
- Fix-ups for small issues in CMakeLists files
- Avoid separate handling of iOS dependencies

Note to the reviewer (@mortenjc): Before this can be merged, a succesful build of this branch must be tested on iOS with Qt6 installed (https://formulae.brew.sh/formula/qt).